*** 3.1.7. Moved to pull_json funciton; modified time of file in manual pull, seems to be working
