+++ .versions
@@ -41,3 +41,4 @@
  *** 3.1.6.4. Trying git pull in git mode...
  *** 3.1.6.5. Saving JSON files into /files/sdss_org_wp_data/
  *** 3.1.6.6. Fixed location of JSON files in sdss_wp_shortcodes.php
+ *** 3.1.7. Moved to pull_json funciton; modified time of file in manual pull, seems to be working
